+++ Magic/CHANGELOG.md
@@ -32,12 +32,13 @@
    - Use `mythic_mob` in a magic mob config to make it build on a mythic mob
    - use `mythic_mob_level` in a config or spawn parameters to set the MM level
 
- - Improved in-game help system `/mhelp` command
+ - Improved in-game help system `/mhelp` command (WIP)
 
  - Add support for JSON chat component messages, escaped with backticks such as
      ```
      &6Swing &7(&b`{"keybind":"key.attack"}`&7) &fyour &a$wand
      ```
+   - Can be used in wand lore, Message action and any in-agme messages.yml text
 
 ## Fixes